complete the modification of the homework modules and problem modules
