+++ oj_backend/oj/account/serializer.py
@@ -1,5 +1,5 @@
 from rest_framework.parsers import JSONParser
-from .models import User,Myclass
+from .models import User
 from rest_framework import serializers
 
 class UserSerialzier(serializers.Serializer):
@@ -13,23 +13,23 @@ class UserSerialzier(serializers.Serializer):
         return User.objects.create(**validated_data)
 
 
-class ShowClassSerializer(serializers.Serializer):
-    class_name=serializers.CharField(max_length=30)
+# class ShowClassSerializer(serializers.Serializer):
+#     class_name=serializers.CharField(max_length=30)
 
 
-class MyclassSerializer(serializers.Serializer):
-    class_name=serializers.CharField(max_length=30)
-    class_admin=serializers.CharField(max_length=30)
-    class_member=serializers.CharField(max_length=30)
-    is_activity=serializers.BooleanField(default=False)
-    create_time=serializers.DateTimeField()
+# class MyclassSerializer(serializers.Serializer):
+#     class_name=serializers.CharField(max_length=30)
+#     class_admin=serializers.CharField(max_length=30)
+#     class_member=serializers.CharField(max_length=30)
+#     is_activity=serializers.BooleanField(default=False)
+#     create_time=serializers.DateTimeField()
 
-    def create(self,validated_data):
-        return Myclass.objects.create(**validated_data)
+#     def create(self,validated_data):
+#         return Myclass.objects.create(**validated_data)
 
-    def update(self,instance,validated_data):
-        instance.class_name=validated_data.get("class_name",instance.class_name)
-        instance.class_admin=validated_data.get("class_admin",instance.class_admin)
-        instance.is_activity=validated_data.get("is_activity",instance.is_activity)
-        instance.save()
-        return instance
+#     def update(self,instance,validated_data):
+#         instance.class_name=validated_data.get("class_name",instance.class_name)
+#         instance.class_admin=validated_data.get("class_admin",instance.class_admin)
+#         instance.is_activity=validated_data.get("is_activity",instance.is_activity)
+#         instance.save()
+#         return instance
