+++ oj_backend/oj/homework/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class HomeworkConfig(AppConfig):
+    name = 'homework'
