+++ oj_backend/oj/submission/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from .models import Submission,Homework_item
+from .models import Submission
 
 class SubmissionSerializer(serializers.Serializer):
     problem_id=serializers.CharField()
@@ -15,19 +15,19 @@ class SubmissionSerializer(serializers.Serializer):
     def create(self,validated_data):
         return Submission.objects.create(**validated_data)
 
-class ClassHomeworkSerializer(serializers.Serializer):
-    class_name=serializers.CharField(max_length=20)
-    homework_item=serializers.CharField(max_length=20)
-    problem_id=serializers.CharField()
-    problem_title=serializers.CharField(max_length=30)
+# class ClassHomeworkSerializer(serializers.Serializer):
+#     class_name=serializers.CharField(max_length=20)
+#     homework_item=serializers.CharField(max_length=20)
+#     problem_id=serializers.CharField()
+#     problem_title=serializers.CharField(max_length=30)
 
-class HomeworkItemSerializer(serializers.Serializer):
-    class_name=serializers.CharField(max_length=20)
-    homework_item=serializers.CharField(max_length=20)
-    homework_item_title=serializers.CharField(max_length=30)
+# class HomeworkItemSerializer(serializers.Serializer):
+#     class_name=serializers.CharField(max_length=20)
+#     homework_item=serializers.CharField(max_length=20)
+#     homework_item_title=serializers.CharField(max_length=30)
 
-    def create(self,validated_data):
-        return Homework_item.objects.create(**validated_data)
+#     def create(self,validated_data):
+#         return Homework_item.objects.create(**validated_data)
