+++ oj_backend/oj/utils/api/utils.py
@@ -14,11 +14,14 @@ class JSONParse(object):
 class MyBaseView(View):
     def _get_request_data(self,request):
         if request.method not in ['GET','DELETE']:
+            
             body=request.body
+            #content_type = request.META.get("CONTENT_TYPE")
             #return HttpResponse(body)
             if body:
                 return JSONParse.parse(body)
-            return {}
+            #return {}
+        
         return request.GET
 
     def dispatch(self,request,*args,**kwargs):
@@ -26,7 +29,10 @@ class MyBaseView(View):
             request.data=self._get_request_data(self.request)
         except ValueError as e:
             return HttpResponse("error")
+        #try:
         return super(MyBaseView,self).dispatch(request,*args,**kwargs)
+        # except Exception as e:
+        #     return HttpResponse(e)
 
 class CSRFExemptMyBaseView(MyBaseView):
     @method_decorator(csrf_exempt)
